Enforce hiraganas for readings.
Depending on the dictionary Mecab has been installed with, it may produce
readings in katakanas or hiraganas. However, readings, in particular when
used as furiganas, are typically displayed as hiraganas. In order to produce
consistant and good readings, this commit converts all the katakanas into
hiraganas for the readings produced by the /furigana and /kana API calls.