Trim furiganas
Rework the whole thing to allow partial furiganas on tokens. This breaks the
XML format and everything. For instance, お母さん is now a single token with
a reading for 母 (I removed CDATA for readability):

    <parse>
      <token>...</token>
      <token>お<reading furigana="かあ">母</reading>さん</token>
    </parse>

Instead of the previous:
    <parse>
      <furigana>...</furigana>
      <furigana>
         <token>お母さん</token>
         <kana>おかあさん</kana>
      </furigana>
    </parse>

This allows to set furiganas more independently from the tokens.

Also grouped the doubtful Kana/Kanji classes into a new Furigana class.